去除无用文件

+++ src/main/java/demo/ms/project/repository/ProjectRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Collection;
 import java.util.List;
 
-public interface ProjectRepository extends JpaRepository<Project,Integer> {
-    List<Project> findByIdIn(Collection<Integer> ids);
+public interface ProjectRepository extends JpaRepository<Project,Long> {
+    List<Project> findByIdIn(Collection<Long> ids);
 }
